Use int64_t instead of long for monomial_word in F4 GB engine
On 32-bit ARM (armhf), long is 32 bits, which caused the F4 Gröbner
basis algorithm to fail with "Internal error: array out of order" and
produce wrong results for EngineTests/GB.Test.LinearAlgebra.m2 over
GF(125) with MonomialSize=>8.  The original code even had a commented-out
int64_t typedef with an "Is this OK?" note acknowledging the uncertainty.

- Extract monomial_word typedef into new f4/monomial-word.hpp so that
  both moninfo.hpp and varpower-monomial.hpp can include it without a
  circular dependency
- Change monomial_word from long to int64_t
- Change varpower_word (ExponentList template parameter) from long to
  monomial_word for consistency
- Update all comparison temporaries from varpower_word to monomial_word
  to avoid narrowing conversions
- Change hash_value return type to uint64_t (consistent with the
  Macaulay2 interpreter; hash values are unsigned fingerprints, not
  monomial slot values)
- Change compare_schreyer tiebreaker parameters to int (they are
  component indices, matching ResMonoid::compare_schreyer)
- Update printf format specifiers to PRId64 where needed
- Behavior is unchanged on 64-bit platforms where long == int64_t

+++ M2/Macaulay2/e/f4/monhashtable.hpp
@@ -12,7 +12,7 @@ class MonomialsWithComponent
 {
  public:
   typedef packed_monomial value;
-  long hash_value(value m) const { return m[0] + m[1]; }
+  uint64_t hash_value(value m) const { return static_cast<uint64_t>(m[0] + m[1]); }
   bool is_equal(value m, value n) const { return mMonoid.is_equal(m, n); }
   void show(value m) const { mMonoid.show(m); }
   MonomialsWithComponent(const MonomialInfo& MI) : mMonoid(MI) {}
@@ -24,7 +24,7 @@ class MonomialsIgnoringComponent
 {
  public:
   typedef packed_monomial value;
-  long hash_value(value m) const { return m[0]; }
+  uint64_t hash_value(value m) const { return static_cast<uint64_t>(m[0]); }
   bool is_equal(value m, value n) const
   {
     return mMonoid.monomial_part_is_equal(m, n);
@@ -68,7 +68,7 @@ class ResMonomialsIgnoringComponent
 // ValueType must implement the following:
 // values should have computed hash values stored with them
 //  typename ValueType::value
-//  long ValueType::hash_value(value m)
+//  uint64_t ValueType::hash_value(value m)
 //  bool ValueType::is_equal(value m, value n)
 //  void ValueType::show(value m) -- prints to stderr
